+++ src/main/java/tenno_mod/powers/DriftingContactPower_TENNO.java
@@ -23,7 +23,7 @@ public class DriftingContactPower_TENNO extends AbstractPower {
     this.amount = amount;
     updateDescription();
     this.type = PowerType.BUFF;
-    this.img = new Texture("imguparrow.png");
+    this.img = new Texture("img/powers/uparrow.png");
   }
 
   public void updateDescription() {